pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.1.2",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-unfold>=0.55.3",
    "djangorestframework>=3.16.0",
    "psycopg>=3.2.7",
    "python-decouple>=3.8",
    "user-agents>=2.2.0",
]
